Fix bug in bond angle calculation
We use integer ids instead of floats.

+++ atooms/postprocessing/ba.py
@@ -87,28 +87,17 @@ class BondAngleDistribution(Correlation):
             system = self.trajectory[i]
             system = change_species(system, 'F')  # species are in fortran style
             side = system.cell.side
-            ids = numpy.array(system.dump('spe'), dtype=numpy.float64)
-            #nn = numpy.ndarray(len(system.particle), dtype=numpy.int32)
-            #neighbors = numpy.ndarray((len(system.particle), 50),
-            #                          dtype=numpy.int32, order='F')
-            #compute.neighbors_list('C', side, self._pos_0[i].transpose(),
-            #                  ids, rcut, nn, neighbors)
+            ids = numpy.array(system.dump('spe'), dtype=numpy.int32)
             nn = numpy.array(0, dtype=numpy.int32)
             neighbors = numpy.ndarray(50, dtype=numpy.int32)
             _log.info('cutoff distance:', rcut)
             for idx in range(len(self._pos_0[i])):
-                # print ids[idx], type(ids[idx])
-                # print ids[idx]-1
-                # print self.rcut[ids[idx]-1: ]
                 compute.neighbors('C', side, self._pos_0[i][idx],
                                   self._pos_1[i].transpose(), ids,
                                   self.rcut[ids[idx]-1, :], nn, neighbors)  # species in Fortran style
                 compute.bond_angle(self._pos_0[i][idx, :], self._pos_1[i].transpose(),
                                    neighbors[0: nn], side,
                                    dtheta, hist_one)
-                #compute.bond_angle(self._pos_0[idx], self._pos_0[i].transpose(),
-                #                   neighbors[idx][0: nn[idx]], side,
-                #                   dtheta, hist_one)
                 hist_all.append(hist_one.copy())
 
         # Normalization
